+++ SimpoleFactory/SimpoleFactory/OperationSub.m
@@ -9,8 +9,8 @@
 #import "OperationSub.h"
 
 @implementation OperationSub
-- (double)result {
-    double result = 0;
+- (NSInteger)result {
+    NSInteger result = 0;
     result = self.numberA - self.numberB;
     return result;
 }
